440: Split voting controller into separate files for each use case
